requirements.txt
fastapi
uvicorn
openai
newspaper3k
beautifulsoup4
twilio
apscheduler
